Updated Coral Scripts
The two coral scripts have been tied together into one (just BaseCoralScript), since I found I could just reference the parent's local scale and script the same things. This also made it easier to tie in the new mechanic where damage to the coral would reduce the scale, as the health and scaling were all being taken care of in the same place.
