--- DIGM530_CoralInvasion/Assets/Scripts/BaseCoralParent.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: a9b3719d26830bd49aa35f772fd5e513
-timeCreated: 1520202840
-licenseType: Free
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
