+++ DIGM530_CoralInvasion/Library/LastSceneManagerSetup.txt
@@ -1,4 +1,4 @@
 sceneSetups:
-- path: Assets/Scenes/ProjectileTest.unity
+- path: Assets/Scenes/LevelBasic.unity
   isLoaded: 1
   isActive: 1
